requirements.txt
garminconnect
schedule
python-dotenv
streamlit
pandas
plotly
psycopg2-binary
fastapi
uvicorn
